aboot: Fix type casting when flashing images to mmc for sparse images

+++ app/aboot/aboot.c
@@ -598,7 +598,7 @@ void cmd_flash_mmc_sparse_img(const char *arg, void *data, unsigned sz)
 	sparse_header_t *sparse_header;
 	chunk_header_t *chunk_header;
 	uint32_t crc32 = 0;
-	unsigned long long total_blocks = 0;
+	uint32_t total_blocks = 0;
 	unsigned long long ptn = 0;
 	unsigned long long size = 0;
 
@@ -660,9 +660,9 @@ void cmd_flash_mmc_sparse_img(const char *arg, void *data, unsigned sz)
 				return;
 			}
 
-			if(mmc_write(ptn + (total_blocks*sparse_header->blk_sz),
-								chunk_data_sz,
-								(unsigned int*)data))
+			if(mmc_write(ptn + ((uint64_t)total_blocks*sparse_header->blk_sz),
+						chunk_data_sz,
+						(unsigned int*)data))
 			{
 				fastboot_fail("flash write failure");
 				return;
